cpu/esp_common: change esp_now key param definition
The 128-bit key for encrypted ESP-NOW communication is defined by a 16-byte array of type `uint8_t [16]` specified by the `ESP_NOW_KEY' configuration parameter in RIOT. By default, this configuration parameter is specified as NULL (no encryption). While in C it is possible to simply define this configuration parameter by something like `#define ESP_NOW_KEY { 0x0f, 0x1e, ... }`, it is not possible in Kconfig to define such a static initializer. Therefore, in regard to the migration to Kconfig, the key is now defined as a string with hex values separated by spaces `#define ESP_NOW_KEY "0f 1e ..."`, which is easy to implement in Kconfig.