pkg/esp32_sdk: use bootloader as RIOT module
Up to now, there were four versions of prebuilt bootloaders for the ESP32 with different settings of logging and color logging. These bootloaders were manually built from the SDK and shipped with RIOT-OS source code.
However there are more settings that affect the bootloader build that are relevant to the app or final board that uses this bootloader. This commit builds the `bootloader.bin` file from the ESP32 SDK source
code. It is built as a module `esp_bootloader` which at the moment doesn't generate any object code for the application and only produces a `bootloader.bin` file set to the `BOOTLOADER_BIN` make variable for the `esptool.inc.mk` to flash.