drivers/at24*: Allow seperate id and eeprom function
The at24mac can have two independent functionalities, eeprom and ids.
Currently bringing in the at24mac brings both the eeprom and id functionality.
This allows the only the eeprom to be selected with at24mac_eeprom.
It also deprecates the at24mac module to eventually only use id functionality.