makefiles/boot/riotboot: set DEBUG_ELFFILE to slot0.elf
When doing a `make debug` on a board with riotboot bootloader, the original
(non-offset) elf file gets selected.

This causes all ROM addresses to be at the wrong offset, leading to strange
debug behavior.

Set `DEBUG_ELFFILE` to the .elf file that already accounts for the bootloader
offset so the debugger gets the correct addresses.