boards/esp32-wrover-kit: revert changes in commit b74c456
In the board definition of `esp32_wrover_kit` default values for `ESP_WIFI_EAP_USER` and `ESP_WIFI_EAP_PASS` had to be defined because this board was used in the CI to compile the optional module `esp_wifi_enterprise`. Now that the CI compilation for the `esp_wifi_enterprise` module is realized by an external board definition `esp32-ci`, these default values should be removed to make the compilation fail if the user did not define these variables.