pkg/emlearn: fix kconfig model
It turns out that #18287 needs to be somewhat reverted.
The make behaviour requires use to add psuedomodules for packages
that are not compiled (ie header only).
The only difference from the revert is that the module is no longer
selected in Kconfig so it will match the modules used list.