rust: Bring lock files in sync with toml files
The riotmodules dependency was missed in bc8ec6d5, and while generally
it does not severly harm builds, it does harm builds riotdocker CI where
branches are switched, which is blocked by dirty files in the checkout.

The riot-wrappers version was missed when what is now 5e75f4bd was
rebased onto fdc4e11a.