core/sched: Provide thread state names for OpenOCD
This provides the thread state names as used in ps() as a symbol that
OpenOCD can read out. This allows implementing OpenOCD's RIOT
awareness in a way that changes to thread states can be picked up at
runtime, without having to patch OpenOCD for every change. Apparently
RIOT is a bit faster moving in this regard than other bare metal RTOSes.