boards: introduce board tiers
The idea is to provide an infrastructure that allows selecting
pre-defined subsets on which to run CI compilation tests. The following
tiers are specified:

- tier 0:
    - only 5 boards, for super fast CI runs
    - based on popularity and heterogeneity of the MCUs used
- tier 1:
    - one board for every MCU family, preferring the MCU variant with
      most peripherals / RAM / flash where possible
- tier 2:
    - all boards

A tier 0 board is also included in tier 1 and tier 2, a tier 1 board
is included in tier 2. All tier are implemented as features. Hence,
a `FEATURES_REQUIRED += board_tier_0` selects only 5 boards, a
`FEATURES_REQUIRED += board_tier_1` selects one board per MCU family
(also including all tier 0 boards), and finally
`FEATURES_REQUIRED += board_tier_2` is fulfilled by all boards.

The idea is to add labels to PRs that will export an `FEATURES_REQUIRED`
environment variable to limit the result of `make info-boards-supported`
to the given set. Maintainers can then decide on the level of CI
coverage a PR should get, depending on how scary the code looks.