cpu/avr*: Fix compilation with GCC 12.2.0
A statement like `LKPR = (1 << CLKPCE);` which for the ATmega328P
translates to `*((volatile uint8_t *)0x61) = (1 << 7);` is considered
as accessing the first array element of a zero sized array by
GCC 12.2.0. Assuming a zero size for memory not allocated by the
compiler is quite insane, so we need to disable the diagnostics for
the affected code. The most targeted approach would be using:

```
#pragma GCC diagnostic push
#pragma GCC diagnostic ignored "-Warray-bounds"
...
#pragma GCC diagnostic pop
```

But adding this around every memory mapped I/O access would render the
code unreadable. Adding this around the `#include <avr/io.h>` instead
would be ideal, but since e.g. `LKPR` is macro, the diagnostic would
be still triggered when that macro is used in RIOT's C code.

Instead, we add `-Wno-array-bounds` to the `CFLAGS`, but only for the
low level AVR/ATmega/ATXmega code. As a result common code can still
profit from the diagnostics, while code working with memory mapped I/O
can still use the AVR libc.