sys/malloc_thread_safe: fix locking if scheduling is not active
If threads are not yet created and scheduling is not active, locking is not required. Even more if the module `core_mutex_priority_inheritance` is used, the mutex must not be locked before threads are created and scheduling is active. Otherwise, priority inheritance will cause a crash if `malloc` is used before threads are created and scheduling is active.