cpu/cortexm_common: disable saving FPU state on IRQ
The context switching code for Cortex M is a bit of a strange in beast
in regard to the FPU state: It will only save it on the stack if the
FPU actually was in use and only restore it if a backup is found on the
stack being switched to.

When switching from a thread that did use the FPU to a thread that did
not yet, the FPU state is consequently not restored and the FPU will
be left in whatever state the previous thread left it. This can have
severe side effects if the new thread starts using the FPU. Especially
with the lazy backup / restore feature of the FPU: If the FPU has saved
registers on the stack (of the old thread) but not yet restored, it
will "restore" it on the next use of the FPU, which corrupts the stack.

Simply disabling backup and restore of the FPU state upon ISR entry
and return avoids this potential memory corruption foot gun. The
trade-off is a FPU state corruption bug whenever an ISR does start
using the FPU.