cpu/stm32/Kconfig: Split clk configs
This PR splits the Kconfig clock configurations for the stm32s
from one file for all series,
into many each series and any common combinations of series.
This makes it more readable and matches what is done in make a bit
closer.

As only the series that is applicable gets sourced, it also removes
a lot of noise created by other series configurations being included.