sys/net/gnrc/sock: fix race condition in timeout handling
`gnrc_sock_recv()` blocks waiting on an `mbox` when receiving. The
timeout implementation just sends a magic timeout message to that mbox
when the timeout timer fires. The implementation assumed that the head
of the mbox will always contain either the received message, or the
magic timeout message, but never both.

This assumption doesn't hold when the timeout fires after the message
was received but before the timer was stopped. This can result in the
subsequent call to `gnrc_sock_recv()` to incorrectly fail with a timeout
right away (without actually waiting), or to a wild pointer
(`_TIMEOUT_MAGIC`) being freed in `gnrc_netreg_unregister()` (whatever
function call comes next).

This fixes the issue by discarding a stale timeout message (if any) from
the mbox while keeping the mbox queue otherwise intact.